move interfaces to model files use Types.ObjectId in interaface as recommended by mongoose 6.x docs
