+++ src/jobQueue/model.ts
@@ -1,6 +1,26 @@
-import { model, Schema } from 'mongoose';
+/**
+ * Job Queues are work queues that deal with an entire job as opposed to processing each lot
+ *  individually.
+ *
+ * This is particularly helpful for stages such as "Planning" or "Shipping", where it is
+ *  more useful to process the entire job once all lots are present, rather than lot by lot.
+ */
+
+import {
+  model, Schema, Document, Types,
+} from 'mongoose';
 import { COLLECTIONS } from '../global.collectionNames';
-import { IJobQueue } from '../global.interfaces';
+
+export interface IJobQueue extends Document {
+  // The name of this queue; e.g. "Planning"
+  name: string;
+
+  // Description of this queue
+  description: string;
+
+  // Jobs currently in this queue
+  jobsInQueue: Types.ObjectId[];
+}
 
 export const JobQueueSchema = new Schema<IJobQueue>({
   name: String,
