+++ src/lot/model.ts
@@ -1,11 +1,30 @@
 /**
- * Temporary schema to get rolling in development.
- * This is fairly representative of what the final schema will look like.
+ * Lots essentially define a bunch of "parts" that move around a site together.
  */
 
-import { model, Schema } from 'mongoose';
+import {
+  model, Schema, Document, Types,
+} from 'mongoose';
 import { COLLECTIONS } from '../global.collectionNames';
-import { ILot } from '../global.interfaces';
+
+export interface ILot extends Document {
+  // Reference to parent Job
+  jobId: Types.ObjectId;
+
+  // is this lot on hold? (overrides job.released)
+  onHold: boolean;
+
+  // is this lot canceled? (overrides job.released)
+  canceled: boolean;
+
+  // quantity to be produced for this lot
+  quantity: number;
+
+  // Revision of this lot (in case of entire quantity scrap)
+  rev: String;
+
+  specialRouter?: Types.ObjectId;
+}
 
 export const LotSchema = new Schema<ILot>({
   jobId: {
@@ -17,6 +36,7 @@ export const LotSchema = new Schema<ILot>({
   canceled: Boolean,
   quantity: Number,
   rev: String,
+
   specialRouter: {
     type: Schema.Types.ObjectId,
     ref: COLLECTIONS.ROUTER,
