+++ src/router/model.ts
@@ -1,6 +1,21 @@
-import { model, Schema } from 'mongoose';
+/**
+ * Routers define the path a job and its lots (or a specific lot) will take once
+ *  a bid has been won.
+ */
+
+import { model, Schema, Document } from 'mongoose';
 import { COLLECTIONS } from '../global.collectionNames';
-import { IRouter } from '../global.interfaces';
+import { IRouteStep } from '../routeStep/model';
+
+export interface IRouterElement {
+  step: IRouteStep;
+  stepCode?: number;
+  stepDetails?: string;
+}
+
+export interface IRouter extends Document {
+  path: IRouterElement[];
+}
 
 export const RouterSchema = new Schema<IRouter>({
   path: [{
