+++ src/site/model.ts
@@ -1,6 +1,34 @@
-import { model, Schema } from 'mongoose';
+/**
+ * Sites are the main hierarchical units of ShopQ.
+ * Each site has its own staff, work queues (job & lot), timezone, location, and name.
+ *
+ * A centralized sales module feeds work to each site.
+ */
+
+import {
+  model, Schema, Document, Types,
+} from 'mongoose';
 import { COLLECTIONS } from '../global.collectionNames';
-import { ISite } from '../global.interfaces';
+
+export interface ISite extends Document {
+  // The name of the site; e.g. HQ
+  name: string;
+
+  // Shipping address of the site; each line is a new string in the array.
+  location: string[];
+
+  // The timezone this site should be synched to.
+  timezone: string;
+
+  // Array of staff members employed at the site.
+  staff: Types.ObjectId;
+
+  // Array of Job Queues.
+  jobQueues: Types.ObjectId[];
+
+  // Array of Lot Queues.
+  lotQueues: Types.ObjectId[];
+}
 
 export const SiteSchema = new Schema<ISite>({
   name: String,
